+++ pineappl/src/lib.rs
@@ -37,6 +37,7 @@
 mod convert;
 
 pub mod bin;
+pub mod boc;
 pub mod empty_subgrid;
 pub mod evolution;
 pub mod fk_table;
@@ -45,7 +46,6 @@ pub mod import_only_subgrid;
 pub mod lagrange_subgrid;
 pub mod lumi;
 pub mod ntuple_subgrid;
-pub mod order;
 pub mod packed_array;
 pub mod pids;
 pub mod sparse_array3;