+++ pineappl_capi/src/lib.rs
@@ -57,9 +57,9 @@
 
 use itertools::izip;
 use pineappl::bin::BinRemapper;
+use pineappl::boc::Order;
 use pineappl::grid::{Grid, GridOptFlags, Ntuple};
 use pineappl::lumi::{LumiCache, LumiEntry};
-use pineappl::order::Order;
 use pineappl::subgrid::{ExtraSubgridParams, SubgridParams};
 use std::collections::HashMap;
 use std::ffi::{CStr, CString};