+++ pineappl_cli/src/import/applgrid.rs
@@ -1,8 +1,8 @@
 use anyhow::Result;
+use pineappl::boc::Order;
 use pineappl::grid::{Convolution, Grid};
 use pineappl::import_only_subgrid::ImportOnlySubgridV2;
 use pineappl::lumi::LumiEntry;
-use pineappl::order::Order;
 use pineappl::sparse_array3::SparseArray3;
 use pineappl::subgrid::{Mu2, SubgridParams};
 use pineappl_applgrid::ffi::{self, grid};