+++ pineappl_cli/src/import/fktable.rs
@@ -1,9 +1,9 @@
 use anyhow::{anyhow, Context, Result};
 use flate2::read::GzDecoder;
+use pineappl::boc::Order;
 use pineappl::grid::{Convolution, Grid};
 use pineappl::import_only_subgrid::ImportOnlySubgridV1;
 use pineappl::lumi_entry;
-use pineappl::order::Order;
 use pineappl::sparse_array3::SparseArray3;
 use pineappl::subgrid::SubgridParams;
 use std::fs::File;